refactor(ari): place `channel` handling under `request_client` module
