+++ arirs/src/request_client/channel/request_params.rs
@@ -193,3 +193,43 @@ impl Serialize for Reason {
         map.end()
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::*;
+
+    #[test]
+    fn serializes_parameters() {
+        let request_client = RequestClient::new("http://localhost:8080/".parse().unwrap(), "asterisk", "asterisk");
+
+        let mut url = request_client.url().join("channel").unwrap();
+
+        request_client.set_authorized_query_params(
+            &mut url,
+            PlayMediaParams {
+                playback_id: None,
+                base_params: PlayMediaBaseParams {
+                    media: &["sound:hello"],
+                    lang: Some("en"),
+                    offset_ms: None,
+                    skip_ms: None,
+                },
+            },
+        );
+
+        let expected = "http://localhost:8080/channel?api_key=asterisk%3Aasterisk&media=sound%3Ahello&lang=en";
+        assert_eq!(expected, url.as_str())
+    }
+
+    #[test]
+    fn serializes_unit_type() {
+        let request_client = RequestClient::new("http://localhost:8080/".parse().unwrap(), "asterisk", "asterisk");
+
+        let mut url = request_client.url().join("channel").unwrap();
+
+        request_client.set_authorized_query_params(&mut url, ());
+
+        let expected = "http://localhost:8080/channel?api_key=asterisk%3Aasterisk";
+        assert_eq!(expected, url.as_str())
+    }
+}
